// Wetax-master/src/view/authenticated/taxReturn/enums.ts
export enum ScreenTypeEnum {
  YesNo = 'YesNo',
  ObjForm = 'ObjForm',
  ArrayForm = 'ArrayForm',
  ArrayOverview = 'ArrayOverview',
  CategoryOverview = 'CategoryOverview',

  ScanOrUploadObj = 'ScanOrUploadObj',
  ScanOrUploadArray = 'ScanOrUploadArray',
  ScanOrUploadArrayBankkonto = 'ScanOrUploadArrayBankkonto',
  GeneratePdf = 'GeneratePdf',
}
// ...
export enum ScreenCategoryEnum {
  Eignung = 'Eignung',
  Einkommen = 'Einkommen',
  Abzuege = 'Abzuege',
  Vermoegen = 'Vermoegen',
  Uebersicht = 'Uebersicht',
}